refactor(mock-test): moved all mocks in src so that they can be accessible for both unit and ui tests
