+++ app/src/main/java/com/android/unio/mocks/firestore/MockReferenceList.kt
@@ -1,5 +1,6 @@
-package com.android.unio.model.firestore
+package com.android.unio.mocks.firestore
 
+import com.android.unio.model.firestore.ReferenceList
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
